+++ app/controllers/users_controller.rb
@@ -59,15 +59,6 @@ class UsersController < ApplicationController
     @user = User.find(params[:id])
   end
 
-  # ログイン済みユーザーかどうか確認
-  def logged_in_user
-    unless logged_in?
-      store_location
-      flash[:danger] = "Please log in."
-      redirect_to login_url
-    end
-  end
-
   # 正しいユーザーかどうか確認
   def correct_user
     @user = User.find(params[:id])
