+++ app/views/users/_follow.html.erb
@@ -1,3 +1,4 @@
+<% # 友達登録済みでもなんか良い感じにdisableのボタンで表現するとそれっぽくなりそう %>
 <%= form_for(current_user.active_relationships.build) do |f| %>
   <div><%= hidden_field_tag :followed_id, @user.id %></div>
   <%= f.submit "Follow", class: "btn btn-primary" %>
