+++ app/views/users/_follow_form.html.erb
@@ -2,6 +2,8 @@
   <div id="follow_form">
   <% if current_user.friend_with?(@user) %>
     <h3>友達です</h3>
+  <% elsif current_user.following?(@user) %>
+    <h3>友達登録を申請中です</h3>
   <% else %>
     <%= render 'follow' %>
   <% end %>
